backend/nonogram-razpoznavalnik/requirements.txt
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
typing-extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0

numpy~=1.21.5
scikit-image~=0.19.1
requests~=2.27.1
Pillow~=9.0.0
scipy~=1.7.3